ビルトインシグネチャ乖離のlint検査を導入し実乖離2件を修正（runtime-builtin-registry 第3段）
docs/design/v0.17.0/runtime-builtin-registry.mdの第3段の実装。
scripts/check_builtin_signatures.pyを新設し、レジストリ表とnative/wasmランタイムC実装の関数シグネチャ（戻り型・引数の幅と個数・可変長）の突き合わせ（297件）とnative/wasm二重実装同士の相互照合（191件）をmake lintとCIのLintジョブへ組み込んだ。
従来この乖離はwasmのcall_indirect検査やABI残留値として実行時にしか現れなかった（N4/V5/M14族）。同名の重複定義も全件検査し、boolのI1/I8表現差は互換、wasmエクスポートでのsize_t等の可変幅型は警告として扱う。
原案の「表からCヘッダを生成してランタイムがinclude」はポインタのpointee型（char*/CmSlice*等）が実装ごとに異なりCの再宣言互換で誤検出になるため、幅正規化（ポインタ→Ptr・スカラ→幅タグ）でのlint時突き合わせへ設計変更した（判断は設計文書の進捗節に記録）。
検査で実乖離2件を検出し修正した:
__builtin_array_findIndex_i64のnative実装がint64_t戻りだった（LLVM宣言・closure変種・wasm実装はi32。ABIの偶然で動作していた）ためi32へ統一。
wasmのcm_format_stringスタブがargc引数を欠いていたためレジストリ宣言（fmt, argc, ...）へ整合。
negative検証: 意図的な不一致（cm_slice_lenのi32戻り定義）でレジストリ照合・相互照合の両方が検出することを確認した。
検証: unit・regression・全バックエンドスイート（interpreter/llvm/llvm-wasm/js/sv/cm-test/libs/i18n/sanitize/ts）通過。
内部リファクタリングのためチュートリアル・VSCode拡張の更新は対象外（構文・キーワードの変更なし）。

+++ src/internal/codegen/llvm/native/runtime_format.c
@@ -1479,7 +1479,8 @@ bool __builtin_array_every_i32(int32_t* arr, int64_t size, bool (*predicate)(int
 }
 
 // findIndex: 条件を満たす最初の要素のインデックス
-int64_t __builtin_array_findIndex_i64(int64_t* arr, int64_t size, bool (*predicate)(int64_t)) {
+// 戻り幅はCmのint（i32）に統一する（LLVM宣言・closure変種・wasm実装と一致。従来のint64_tはABIの偶然で動作していた）
+int32_t __builtin_array_findIndex_i64(int64_t* arr, int64_t size, bool (*predicate)(int64_t)) {
     CM_HOF_UNWRAP(arr, size);
     if (!arr || !predicate)
         return -1;
